requirements.txt
anyio==3.6.1
asyncpg==0.26.0
black==22.10.0
click==8.1.3
fastapi==0.85.0
graphql-core==3.2.3
h11==0.14.0
httptools==0.5.0
idna==3.4
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
psycopg==3.1.3
psycopg-c==3.1.3
psycopg-pool==3.1.2
pydantic==1.10.2
Pygments==2.13.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
strawberry-graphql==0.133.5
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.18.3
uvloop==0.17.0
watchfiles==0.17.0
websockets==10.3
